+++ Procfile
@@ -1,0 +1,3 @@
+server:    bundle exec rails s
+delayed: bundle exec ./script/delayed_job run
+warren:  bundle exec warren consumer start
