+++ config/application.rb
@@ -75,8 +75,6 @@ module Sequencescape
     # Run "rake -D time" for a list of tasks for finding time zone names. Uncomment to use default local time.
     config.time_zone = 'London'
 
-    config.warren = config_for(:warren)
-
     # Your secret key for verifying cookie session data integrity.
     # If you change this key, all old sessions will become invalid!
     # Make sure the secret is at least 30 characters and all random,
