+++ script/delayed_job
@@ -1,5 +1,4 @@
 #!/usr/bin/env ruby
-
 require File.expand_path(File.join(File.dirname(__dir__), 'config', 'environment'))
 require 'delayed/command'
 Delayed::Command.new(ARGV).daemonize
